Add possibility to not store all the samples that leak memory
It affects the possibility to generate report that requires data.
So, add new config option called '--generate-report' to cover the
situation with memory leaks.

If it is set then we store all the samples as before and generate
final report based on the stored data. Side-effect - linear memory
leaks.
If it is not set (default) then the samples data won't be stored
and final report won't be generated keeping stable memory consumption.