Make interrupting by Ctrl-C more reliable
The retrying logic made Ctrl-C handling a bit broken,
because the streams could be terminated only at the
complete cycles. If a query retries for a long time
it prevented terminating the program.

This commit implements interruption logic in
a different way, by using tokio::signal
and tokio::select! so all streams are terminated
immediately, regardless of their progress.