Improved sampling
Previously switching to the next sample could only
happen at the moment of completing a cycle. If the db
didn't respond for a longer period of time, the samples
collected were of unequal sizes or weren't recorded at all.

Now the timer for sampling (chunking) ticks independently
of the workload progress.