Fix blob function usage
In the PR [1] with the switch from 'rune-0.12' to 'rune-0.13'
the 'blob' function changed it's return type from 'rune::runtime::Bytes'
to the 'Vec<u8>'.
Such a change required also an update to the input values parser
to expect one more input type match to the 'ColumnType::Blob'.

So, add the missing parsing element to make the 'blob' function work
again as it was before.
It will allow to avoid manual wrapping into the
'rune::runtime::Bytes' type in rune scripts.

[1] https://github.com/pkolaczk/latte/pull/96